refactor: delete screens package, move files to root
