+++ examples/llm_inference/android/app/src/main/java/com/example/mediapipe/llminference/ChatUiState.kt
@@ -1,4 +1,4 @@
-package com.example.mediapipe.llminference.ui.screens
+package com.example.mediapipe.llminference
 
 import androidx.compose.runtime.toMutableStateList
