+++ examples/llm_inference/android/app/src/main/java/com/example/mediapipe/llminference/InferenceModel.kt
@@ -33,7 +33,7 @@ class InferenceModel private constructor(context: Context) {
     }
 
     companion object {
-        private const val MODEL_PATH = "/data/local/tmp/g_mini_2b_pt_final_gpu/model_ml_drift.tflite"
+        private const val MODEL_PATH = "/data/local/tmp/g_mini_2b_pt_final_gpu/model_ml_drift2.tflite"
         private var instance: InferenceModel? = null
 
         fun getInstance(context: Context): InferenceModel {
